+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Release Notes
 
+## Next
+- fix(android): Add `getTotalDiskCapacityNew` and `getFreeDiskStorageNew` - and their sync variants. Calculations now take into account both `Root` and `Data` file systems.
+
 ## 5.6.3
 
 - fix(ios): fix WkWebView crash from parallel getUserAgent calls (#1050, thanks @RojoHub!)